--- src/components/mobile-menu/collapse-items/OptionsCollapse.vue
@@ -1,17 +1,0 @@
-<template>
-  <NCollapseItem title="Options">
-    <OptionsMenu></OptionsMenu>
-  </NCollapseItem>
-</template>
-
-<script setup>
-import { NCollapseItem } from 'naive-ui'
-
-import OptionsMenu from '@/components/mobile-menu/OptionsMenu.vue'
-
-// import { state } from '@/static/state'
-
-</script>
-
-<style lang="less">
-</style>
