+++ src/components/settings/SettingsOverlay.vue
@@ -1,10 +1,19 @@
 <template>
   <div class="hud-settings-overlay" @click="onCloseModal()">
     <div class="hud-settings-container" @click="$event.stopPropagation()">
-      <MapResize v-if="state.options.showSideMap" />
-      <InterfaceToggles />
-      <FontSelector />
-      <SettingsActions />
+      <div class="row">
+        <InterfaceToggles />
+        <FontSelector />
+        <MusicPlayer />
+        <MapResize v-if="state.options.showSideMap" />
+        <GamepadSettings />
+      </div>
+      <div class="row">
+        <SettingsActions />
+      </div>
+      <div class="row" v-if="state.showMoreOptions">
+        <OptionToggles />
+      </div>
     </div>
   </div>
 </template>
@@ -16,6 +25,9 @@ import FontSelector from '@/components/settings/FontSelector.vue'
 import MapResize from '@/components/settings/MapResize.vue'
 import InterfaceToggles from '@/components/settings/InterfaceToggles.vue'
 import SettingsActions from '@/components/settings/SettingsActions.vue'
+import MusicPlayer from '@/components/common/MusicPlayer.vue'
+import OptionToggles from '@/components/settings/OptionToggles.vue'
+import GamepadSettings from '@/components/settings/GamepadSettings.vue'
 
 function onCloseModal () {
   if (!state.modals.settingsModal) {
@@ -42,7 +54,7 @@ onBeforeUnmount(() => {
   left: 0;
   width: 100%;
   height: 100%;
-  z-index: 2;
+  z-index: 30;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -57,13 +69,23 @@ onBeforeUnmount(() => {
     display: flex;
     flex-direction: row;
     justify-content: center;
-    align-items: center;
+    align-items: flex-start;
     flex-wrap: nowrap;
     padding: 5px;
     gap: 20px;
     background-color: rgb(27 38 45 / 90%);
     border: 1px solid rgb(69 100 119);
     border-radius: 4px;
+    max-height: calc(100vh - 100px);
+    overflow-y: auto;
+
+    .row {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+      gap: 10px;
+    }
 
     @media screen and (max-width: 750px) {
       // top: 0;
